pyproject.toml
[project]
name = "uc-intg-monoprice-htp1"
dynamic = ["version"]
description = "Monoprice HTP-1 AV Receiver integration for Unfolded Circle Remote"
authors = [
    {name = "Meir Miyara", email = "meir.miyara@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MPL-2.0"}
keywords = ["unfoldedcircle", "integration", "monoprice", "htp1", "receiver"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL-2.0)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "ucapi-framework>=1.4.0",
    "aiohttp>=3.9.0",
]

[project.urls]
Homepage = "https://github.com/mase1981/uc-intg-monoprice-htp1"
Repository = "https://github.com/mase1981/uc-intg-monoprice-htp1"
Issues = "https://github.com/mase1981/uc-intg-monoprice-htp1/issues"

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["intg_monoprice_htp1"]

[tool.setuptools.dynamic]
version = {attr = "intg_monoprice_htp1.__version__"}
